Fixed broken typecast (would have cast Double to Long; boom).
git-svn-id: https://svn.apache.org/repos/asf/struts/struts2/trunk@595360 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/struts2/views/jsp/iterator/IteratorGeneratorTag.java
@@ -207,7 +207,7 @@ public class IteratorGeneratorTag extends StrutsBodyTagSupport {
                 count = ((Long)countObj).intValue();
             }
             else if (countObj instanceof Double) {
-                count = ((Long)countObj).intValue();
+                count = ((Double) countObj).intValue();
             }
             else if (countObj instanceof String) {
                 try {
